+++ tests/ops/fixtures/application_fixtures.py
@@ -1060,6 +1060,7 @@ def example_datasets() -> List[Dict]:
         "data/dataset/mariadb_example_test_dataset.yml",
         "data/dataset/bigquery_example_test_dataset.yml",
         "data/dataset/manual_dataset.yml",
+        "data/dataset/email_dataset.yml",
     ]
     for filename in example_filenames:
         example_datasets += load_dataset(filename)